requirements.txt
opencv-python>=4.5.0
fast_alpr
onnxruntime
pytesseract
easyocr>=1.6.0
Pillow>=8.0.0